Separate code into components, update description save method
